web_ui/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
channels>=4.0.0
channels-redis>=4.1.0
psycopg2-binary>=2.9.0
mysqlclient>=2.2.0
redis>=4.0.0
celery>=5.3.0
httpx
paramiko
django-cors-headers>=4.0.0
django-filter>=23.0
Pillow>=10.0.0
python-decouple>=3.8
cryptography>=41.0.0
pyjwt>=2.8.0
requests>=2.31.0
websockets>=11.0
